refactor(frontend): Home(예약 하기) 리팩토링
